fix: improve accessibility on mobile
Fixes #6.

To be honest I did not think about this at all. I simply copied the code from the original »Bear Blog«

https://github.com/HermanMartinus/bearblog/blob/master/templates/base.html#L7

It seems like iOS ignores the declarations of user-scalable, minimum-scale, and maximum-scale, as of iOS 10, see https://sitebulb.com/hints/mobile-friendly/the-viewport-meta-tag-prevents-the-user-from-scaling/ so we might as well remove this.