Fix a remaining issue with b8c8871 (#20)
It turns out my original proposal for #20 was correct, but not for the
reasons I thought. We need to set both `background-color` and `color` to
`initial` on `div.highlight pre` because that's where Chroma sets those
colors (including the default `color` if configured). Setting to
`initial` there makes it so that if the selected style *doesn't*
configure a default color, we'll use the `initial` color, which is going
to be the right choice because the reason style author left that unset
because they didn't consider dark color schemes messing with their
style. Then we `unset` the colors on `div.highlight code` because
otherwise the `code` colors from the theme will override the colors that
would otherwise be inherited from Chroma's `<pre>` element.

Stricly speaking I can't say that setting `background-color` to
`initial` is required, because I haven't found a Chroma style that
*doesn't* set a `background-color`, but I figure it's possible (at least
for a light theme) and it makes sense to fix it just in case, and causes
no harm otherwise.