Use hugo.toml instead of config.toml
Use the file hugo.toml in both the README.md and the exampleSite, since
Hugo will look for this file first.

The current instructions (using config.toml) in a new Hugo site
(that has a hugo.toml-file) means that the site doesn't work, or the
theme is not activated.

Note that this is a breaking change for sites running on older versions.
(But v0.110.0 was released almost a year ago.)

Relevant Hugo commit:

https://github.com/gohugoio/hugo/commit/f38a2fbd2e4de7f095a833b448cb8bc053955ce2